Moved JobRepository filters to subfolder
